Adding missing {} around a logic check; bug #64514.
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/genkernel/trunk@131 67a159dc-881f-0410-a524-ba9dfbe2cb84

+++ gen_package.sh
@@ -7,7 +7,7 @@ gen_minkernpackage()
 	mkdir "${TEMP}/minkernpackage" || gen_die 'Could not make a directory for the kernel package!'
 	cd "${KERNEL_DIR}"
 	cp "${KERNEL_BINARY}" "${TEMP}/minkernpackage/kernel-${KV}" || gen_die 'Could not the copy kernel for the kernel package!'
-	[ "${BUILD_INITRD}" -ne 0 ] && cp "${TEMP}/initrd-${KV}" "${TEMP}/minkernpackage/initrd-${KV}" || gen_die 'Could not copy the initrd for the kernel package!'
+	[ "${BUILD_INITRD}" -ne 0 ] && { cp "${TEMP}/initrd-${KV}" "${TEMP}/minkernpackage/initrd-${KV}" || gen_die 'Could not copy the initrd for the kernel package!'; }
 	cd "${TEMP}/minkernpackage" 
 	tar -jcpf ${MINKERNPACKAGE} * || gen_die 'Could not compress the kernel package!'
 	cd "${TEMP}" && rm -rf "${TEMP}/minkernpackage" > /dev/null 2>&1
